chore(classes): move classes into separate files and change some test implementations
