+++ src/lib/FetchClientResponse.ts
@@ -1,0 +1,55 @@
+
+import {IFetchResponse} from "../main";
+
+export class FetchClientResponse<T> {
+    fetchResponse: IFetchResponse;
+    body: any;
+
+    get hasError(): boolean {
+        return !this.fetchResponse.ok;
+    }
+
+    get status() {
+        return this.fetchResponse.status;
+    }
+
+    get statusText() {
+        return this.fetchResponse.statusText;
+    }
+
+    get contentType() {
+        if (!this.fetchResponse.headers.has("Content-Type")) {
+            return undefined;
+        }
+        return this.fetchResponse.headers.get("Content-Type").split(";")[0];
+    }
+
+    get ok() {
+        return this.fetchResponse.ok;
+    }
+
+    get hasData(): boolean {
+        return !(this.status === 204);
+    }
+
+    get data(): T {
+        switch (this.contentType) {
+            case "application/json":
+                try {
+                    return JSON.parse(this.body) as T;
+                } catch (e) {
+                    throw Error("Response body can't be parsed.");
+                }
+            case undefined:
+            case "":
+                return undefined;
+        }
+
+        throw new Error(`Unknown Content-Type "${this.contentType}"`);
+    }
+
+    constructor(fetchResponse: IFetchResponse, body: any) {
+        this.fetchResponse = fetchResponse;
+        this.body = body;
+    }
+}
